// ui/app/src/model/folder.ts
// Copyright The Perses Authors
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
// http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

import { ProjectMetadata } from '@perses-dev/client';

export interface FolderDisplay {
  name?: string;
}

export interface FolderSpec {
  display?: FolderDisplay;
  items?: FolderItem[];
}

export interface FolderItem {
  kind: 'Folder' | 'Dashboard';
  name: string;
  items?: FolderItem[];
}

export interface FolderResource {
  kind: 'Folder';
  metadata: ProjectMetadata;
  spec: FolderSpec;
}